MAINT: moved the `nested_getstate` to workflow_writer and `pop_*` methods
