+++ force_bdss/io/workflow_writer.py
@@ -131,3 +131,20 @@ def pop_dunder_recursive(dictionary):
                     pop_dunder_recursive(element)
 
     return dictionary
+
+
+def nested_getstate(state_dict):
+    # We safely attempt to get the state of the nested objects in `state_dict`
+    # on the zero and first levels using __getstate__.
+    # If the `state_dict` item is an iterable, we __getstate__ of the iterable
+    # elements. Otherwise, we __getstate__ the item itself.
+    # If we can't __getstate__ of the item, we leave it as it is.
+    for key in state_dict:
+        try:
+            if isinstance(state_dict[key], (tuple, list)):
+                state_dict[key] = [el.__getstate__() for el in state_dict[key]]
+            else:
+                state_dict[key] = state_dict[key].__getstate__()
+        except AttributeError:
+            pass
+    return state_dict
